+++ docker.go
@@ -15,6 +15,7 @@ import (
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/client"
+	"github.com/docker/docker/pkg/stdcopy"
 	"golang.org/x/crypto/ssh/terminal"
 	"golang.org/x/net/context"
 	pb "gopkg.in/cheggaaa/pb.v1"
@@ -219,6 +220,11 @@ func (c *DockerClient) StartContainer(rm bool, name string) (string, error) {
 			return "", fmt.Errorf("Failed to fetch image: %s", err)
 		}
 	}
+
+	// Disable TTY if we're non-interactive, or if we're not in a terminal
+	if nonInteractive || !terminal.IsTerminal(int(os.Stdout.Fd())) {
+		c.Conf.Tty = false
+	}
 	resp, err := c.Cli.ContainerCreate(context.Background(), c.Conf, c.HostConf, c.NetConf, name)
 
 	if err != nil {
@@ -237,6 +243,11 @@ func (c *DockerClient) StartContainer(rm bool, name string) (string, error) {
 
 	// Set the TTY size to match the host terminal
 	fd := int(os.Stdin.Fd())
+
+	// TODO: I would argue that, at some point, we change this behaviour to be more inline with Docker:
+	// -i interactive - default true if in a terminal, or if we're piping stuff in with stdin
+	// -t tty         - default true if in a terminal
+	// But given Cali doesn't currently support piping stuff in from stdin... not important yet
 
 	if !nonInteractive && terminal.IsTerminal(int(os.Stdout.Fd())) {
 		// While we have a container running, create a buffer for the pscli logs
@@ -332,7 +343,7 @@ func (c *DockerClient) startContainerNonInteractive(containerID string) error {
 		return fmt.Errorf("Failed to get container logs: %s", err)
 	}
 
-	_, err = io.Copy(os.Stdout, containerLogs)
+	_, err = stdcopy.StdCopy(os.Stdout, os.Stderr, containerLogs)
 	if err != nil {
 		return fmt.Errorf("Failed to get container logs: %s", err)
 	}