+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DockerClient.InitDocker does nothing if Docker client was previously initialised
 - All DockerClient methods which rely on an initialised Docker client now run InitDocker. As a result, a few DockerClient methods now have the possibilities of returning errors: ContainerExists, ImageExists
 - Auto-detects whether host process is a terminal, and uses that to automatically decide whether to run container non-interactive
+- If running non-interactive, disable TTY on container
 
 [0.1.2]:      https://github.com/skybet/cali/compare/v0.1.1...v0.1.2
 ## [0.1.2] - 2018-04-07